requirement.txt
pillow==10.3.0
streamlit==1.34.0
opencv-python==4.10.0.84
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.16.0
